Add a reservation ID to vm_map_entry_t.
To avoid adding a reservation allocator for now, set the reservation
ID to the first address of each new-mapping (or set of mappings in the
case of stacks+guards and aout binaries).

Forbid coalescing maps with different reservation IDs.  This has the
unfortunate side-effect to not coalescing heap mappings, but is
necessicary if we're going to allow revocation of pointeres returned by
mmap.